Allow installing as a python package directly Move requirements.txt to pyproject.toml Use a proper module structure
