+++ Dockerfile
@@ -16,10 +16,10 @@
 FROM python:3-slim-buster
 
 RUN mkdir /home/sa_scanner
-COPY *.py /home/sa_scanner/
-COPY requirements.txt /home/sa_scanner/
+COPY src/ /home/sa_scanner/
+COPY pyproject.toml /home/sa_scanner/
 
-RUN pip install -r /home/sa_scanner/requirements.txt
+RUN pip install /home/sa_scanner
 
 WORKDIR /home/sa_scanner/
-ENTRYPOINT ["/usr/local/bin/python3", "/home/sa_scanner"]
+ENTRYPOINT ["gcp-scanner"]
