+++ pyproject.toml
@@ -1,0 +1,25 @@
+[build-system]
+requires = ["setuptools"]
+
+[project]
+name = "gcp_scanner"
+description = "Google GCP Scanner"
+readme = "README.md"
+requires-python = ">=3.7"
+keywords = ["gcp"]
+license = {text = "Apache License"}
+classifiers = [
+    "Programming Language :: Python :: 3",
+]
+dependencies = [
+    "pyu2f==0.1.5",
+    "google-api-python-client==2.9.0",
+    "google-cloud-container==2.4.1",
+    "google-cloud-iam==2.2.0",
+    "httplib2==0.19.1",
+    "requests==2.25.1"
+]
+dynamic = ["version"]
+
+[project.scripts]
+gcp-scanner = "gcp_scanner.scanner:main"
