+++ src/gcp_scanner/test/cloudstore_instances
