+++ src/gcp_scanner/test/compute_images
