+++ src/gcp_scanner/test/gke_images
