+++ src/gcp_scanner/test/iam_policy
