+++ src/gcp_scanner/test/managed_zones
