+++ src/gcp_scanner/test/static_ips
